# openapi.yaml
openapi: 3.0.2
info:
  description: OFAC (Office of Foreign Assets Control) API is designed to facilitate the enforcement of US government economic sanctions programs required by federal law. This project implements a modern REST HTTP API for companies and organizations to obey federal law and use OFAC data in their applications.
  version: v1
  title: OFAC API
  contact:
    url: https://github.com/moov-io/ofac
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8084
    description: Local development
...
paths:
...
  /downloads:
    get:
      tags:
        - OFAC
      summary: Return list of recent downloads of OFAC data
      operationId: getLatestDownloads
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            example: 25
          description: Maximum results returned by a search
      responses:
        '200':
          description: Recent timestamps and counts of parsed objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Downloads'
components:
  schemas:
...
    Downloads:
      type: array
      items:
        $ref: '#/components/schemas/Download'
    Download:
      description: Metadata and stats about downloaded OFAC data
      properties:
        SDNs:
          type: integer
          example: 7414
        altNames:
          type: integer
          example: 9729
        addresses:
          type: integer
          example: 11747
        timestamp:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
...